refactor tab title
